+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.0
 * String.ToValueOr(...) fixes: properly support TimeSpan, support Enums, more test coverage
-* Assembly.GetResourceBySuffixAs[Stream|String] implementations
+* Assembly.GetResourceBySuffixAs[Stream|String] extension methods
+* [Memory|Local]ScopedTempFile[Manager] implementations
+* DirectoryInfo.GetTempFileName(...) extension method
 
 ## 0.1.0
 * Integrated GitVer for SemVer management